requirements.txt
python_dotenv==0.6.3
feedparser==5.2.1
httplib2==0.9.2
pytz==2016.7
requests==2.9.1
google_api_python_client==1.6.2
